Update instructions for running tests
I tried using the testing instructions but they do not work as they
depend on a bunch of test dependencies being in place. While
investigating I also discovered that running tests through setuptools is
[no longer recommended](https://docs.pytest.org/en/latest/explanation/goodpractices.html#do-not-run-via-setuptools).
I have updated the instructions to follow the steps taken in the GitHub
action.

I have also explicitly set the `asyncio_mode` as this was raising a warning.
It looks like the other warnings will be addressed [when asyncssh is
updated](https://github.com/ronf/asyncssh/issues/473).